printing

+++ instruction.h
@@ -71,7 +71,7 @@ class input_instr {
     // constant offset between memory read and value stored in reg
     long long int offset1;
     long long int offset2;
-    
+
     // instruction pointer or PC (Program Counter)
     uint64_t ip;
 
@@ -297,8 +297,8 @@ class ooo_model_instr {
     // cout << "*** " << instr_id << " ***" << endl;
     cout << hex << "ip: 0x" << (uint64_t)ip << dec << endl;
     cout << hex << "op: 0x" << (uint64_t)op << dec << endl;
-    cout << "first offset: " << (uint64_t)offset1 << endl;
-    cout << "second offset: " << (uint64_t)offset2 << endl;
+    cout << "first offset: " << offset1 << endl;
+    cout << "second offset: " << offset2 << endl;
     cout << (uint32_t)is_branch << " " << (uint32_t)branch_taken << endl;
     cout << "source registers:";
     for(uint32_t i=0; i<NUM_INSTR_SOURCES; i++)
